Cast divisor to double to avoid integer division

+++ clients/src/main/java/org/apache/kafka/clients/ClusterConnectionStates.java
@@ -34,7 +34,7 @@ final class ClusterConnectionStates {
     public ClusterConnectionStates(long reconnectBackoffMs, long reconnectBackoffMax) {
         this.reconnectBackoffMs = reconnectBackoffMs;
         this.reconnectBackoffMax = reconnectBackoffMax;
-        this.reconnectBackoffMaxExp = Math.log(reconnectBackoffMax / Math.max(reconnectBackoffMs, 1)) / Math.log(RECONNECT_BACKOFF_EXP_BASE);
+        this.reconnectBackoffMaxExp = Math.log(reconnectBackoffMax / (double) Math.max(reconnectBackoffMs, 1)) / Math.log(RECONNECT_BACKOFF_EXP_BASE);
         this.nodeState = new HashMap<String, NodeConnectionState>();
     }
